usb: gadget: storage: Prevent race condition from stalling thread
Parts of the mass storage gadget driver are prone to race conditions
which could leave the main worker thread in an unwakeable state.
This will lead to the function to not be responsive to the host's
command requests and will result in a host-issued reset. Protect
concurrent accesses to certain state variables via spinlock to
prevent this from happening.

CRs-Fixed: 480025
Change-Id: I5232d9367ff311817f63f8bfbea0c78ef2b5a4db
Signed-off-by: Jack Pham <jackp@codeaurora.org>