power: workaround USB disconnect while flash enabled
When SMBB_BOOST is enabled the FSM turns off the
USB OVP FET using a signal which in turn disables
the USB input comparators, thus causing the usbin_valid
signal to go low and causing a disconnect.

Fix this by allowing the Flash/Torch driver to use the
pm8941_smbb_boost regulator to software enable the boost
and subsequently trigger the software workaround which
overrides the USB comparator output and CHG_GONE outputs.

CRs-Fixed: 473004
Change-Id: I76ea029d0c1c96e6d9adf3268c5ab18aaa52be60
Signed-off-by: David Keitel <dkeitel@codeaurora.org>