thermal: tsens8974: Use a dedicated tsense high-priority workqueue
Thermal-mitigation software subscribing the temperature updates
relies upon timely notifications of temperature changes. The existing
implementation scheduled these notifications on the system workqueue,
which may result in addition latency when the system is under heavy
load. Use a high-priority workqueue to solve this problem and ensure
time-sensitive temperature notifications are received as quickly as
possible.

Change-Id: Ic8a18cecfae5a26d61d7179d9d9a2a1273e72744
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>