ASoC: 8974: remove unsed clock API
MSM8974 uses DIV_CLK1 from PMIC 8941 to supply master clock
to Audio codec. This clock is set to freq of 9.6MHZ by
default and cannot be set to any other frequency as it is
exposed by RPM.
The function clk_set_rate, if used, will always fail and
will return -1 error. Remove the function to avoid any kind
of confusion and as it is redundant in terms of functionality.
For clocks exposed by RPM, clk_prepare_enable and
clk_disable_unprepare are enough

CRs-fixed: 493045
Change-Id: I06868297e3cda1f3a9b7a8e36761d852c916d9c1
Signed-off-by: Banajit Goswami <bgoswa@codeaurora.org>