msm: acpuclock-9625: Use 300MHz for cpu power collapse
If acpu power collapse is taking a long time, it will result in
a USB device compliance test failure due to not satisfy the USB
spec. Therefore, increase the power collapse frequency to address
this issue.

In addition, prepare and enable the always on source gpll0 for
the 300MHz so that the source is always prepared during power
collapse.

Change-Id: Ic67529fde291da7c15035cb6aef1dcf0e178d9a6
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>