USB: dwc3: msm: Perform SSPHY register read twice
CR protocol interface is used for reading and writing SSPHY registers.
CR protocol may be performed incorrectly since both the acknowledge
and data are synchronized, and the exact delay of the synchronizers
is uncertain. Due to this hardware bug, first read of SSPHY register
may be incorrect. Hence SW should perform SSPHY register read twice
and use only second read. This is not applicable to SSPHY register
write.

CRs-Fixed: 499129
Change-Id: I5d17cf2f6da997845fa63460d3159213d6bb24ac
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>