power: qpnp-bms: Store FCC learnt data in BMS registers
BMS module supports 16 (8-bit) battery backed-up registers.
Use these registers to store the learnt FCC data. Prior to
this change, the generated FCC data was expected to be
maintained by userspace.

Every new FCC value is first interpolated to a default
temperature (this eliminates the use of storing the
temperature at which the FCC is generated). The FCC value
(in mAh) is mapped to a 8-bit value and stored accordingly.
Same encoding mechanism is used to store the charge-cycles.

Change-Id: I1f5870afbccb306cc2e54b1d539f8446c10540f6
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>