diag: Fix support for resizing buffers used in reading smd data
When an oversized packet is being read from an smd
channel we resize the buffer into which the data is being
read. The current implementation does not correctly use
and reassign the resized buffer. Make modifications to correctly
reassign the resized buffer so that it can be used properly.

CRs-Fixed: 532167
Change-Id: I1031b80147c90193c358ab3c32a3c0fedae5348e
Signed-off-by: Dixon Peterson <dixonp@codeaurora.org>