ASoC: msm8226: Add AANC support for msm8226
Initiate sending commands to DSP for AANC. DSP will use
the information from the commands to know which codec
register to write anc gain which is determined by the
AANC algorithm.
Resend the commands after SubSystem Restart, because
DSP will lose the information and will not know
what codec register to write to or which AANC version
to use.

Change-Id: I442b930efd091a82e90cecece0edd5ac38815ddc
Signed-off-by: Damir Didjusto <damird@codeaurora.org>