msm: mdss: wait for the first display update
Display backlight is enabled after commit kickoff function
returns.  However, the return of the function does not mean
the data has been sent to the panel yet.  During suspend and
resume use case, it results a split of second of screen
corruption.  This makes sure backlight is enabled only after
the first display update is complete.

Change-Id: Id5e6bada69c51180ebedfdb301d4d68fc12877b0
Signed-off-by: Xiaoming Zhou <zhoux@codeaurora.org>