ARM: dts: msm: remove rpm-standalone flag for 8610
The rpm-standalone flag is used to succeed all rpm
communication from PM driver incase of SMD channel
fails to open. This is to allow drivers to succeed
in a pre-silicon environment where SMD and RPM drivers
not fully functional until the actual silicon. After
actual silicon bring up it is not required to have
this flag in DT. Hence removing the flag.

CRs-fixed: 562322
Change-Id: I0837977f027f2926643176970b22813490241fcd
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>